Improve error handling
The main problem with Box<dyn Error> is that it's not thread-safe, which
is my main motivation for making this change.

Instead, we can use the `anyhow` crate that effectively behaves the same
but is thread-safe and always provides backtraces.

We could use generic anyhow:Errors everywhere but I think having proper
structured errors is cleaner, and as I discovered when writing ApiError,
reduces duplication. And it paves the way for providing detailed errors
of our own instead of using anyhow in the future.

Most functions now return Result<A, anyhow::Error> except for a few
where we can return a more specific type of error like ApiError or
PageError.

For the most part this should be backwards-compatble since anyhow::Error
implements std::error::Error but there are a few functions that used to
return Result<A, &str>, which now return proper Errors.